add better tooling for asset management
