+++ mani_skill/utils/assets/README.md
@@ -1,0 +1,4 @@
+# ManiSkill Asset Management
+
+Code for asset management in ManiSkill
+
