+++ mani_skill/utils/assets/__init__.py
@@ -1,0 +1,1 @@
+from .data import DATA_GROUPS, DATA_SOURCES, DataSource, is_data_source_downloaded
